Updated Read_Master to take folder inputs from an external source to allow for the a folder name to be passed to the function later.
PSA_Main_UI updated to perform a read Master call with default values when the UI starts. The Read master now populates values in a single line.

I am aiming to get the UI to open a file dialog when the anlayser number is changed in the combo box. This will generate the folder name and allow the calibrator to switch between analsyers more easily.

I am also trying to get the home page to activate and update automatically. It currently only updates the analyser when you actually change an analyser number