Update on 22 March 2022
Detecotr Stability table changed from 3 tables to a single wide table showing the status of all 16 detectors. This was in response to the QTableWidget automatically indexing to the next row when more thn the number of columns worth of data was available.
This could probably be fixed by controlling the loading of the data more acurately. At this stage i couldn;t be bothered since it is the UI anyway. the report is what is really going to matter and as long as we can get at the data the report can be built as required.

need to change the Temperature stability table to drop downs and highlight them in yellow when not filled and change them to white when filled in