Update on 24 March 2022
Updated page 1 with similar changes as implemented for the temperature stability table. The required baxoes turn grey after an initial change is made.

Updated all pages to remove the regnerate report button and changed page control to be a simple close which returns you to the home page where you can then access the next page you are looking for directly.