Update on 28 March 2022
Introduced function which will dump data to a JSON file when the page control button hides the particular page.

unfortunately, not all the UI widgets are correctly labelled. This needs to be sorted out and the UI regenerated before we can complete the Data interchange file.

We will commit and merge at this stage and then rebase the branch and continue after the UI has been sorted out.