Update on 29 March 2022 Lunchtime
Progress made on the JSON. All page 1 data is being pulled into the JSON using loops.
Page 2 detector stability data and combobox data also being pulled into JSON.

Now working on how to encode the plots but this is not critical at this stage. We can send these as part of a zip file as required.