Update on 4 April 2022 14:13
Only current info is dumping to JSON at the moment. This is due to the Table not being populated with historical Data.

There is no mechanism to load historical data at the moment. this will have to be raised as an issue and includeed in a different commit. Currently the code for pulling the data from these items into the JSON is in the code but commented waiting for historical data to be available.

UpdatePg6 function is not updating the table correctly. There are a few weird things going on there. This needs to be resolved. This will be dealt with in the next commit.