Update on 4 April 2022 16:10
Page 6 data is updating current data correctly now. Will have to debug over an extended period but for now it is doing what it is supposed to do.

The JSON is also updating the way it is supposed to.
Added bank cell creation for teh comment cells to the Page6Update function so that Comment cells are created blank. this may lead to an issue later but we will have to see if this comes back in the debug reporting.