Update on 6 April 16:30
Service Engineer Code is now decoding and generating the images.
This versiion is also looking for a ReportData.json file in a specified location and looking for a resources folder in that location as well. It then places the images into that resources location.

I am not sure if a should stick to this or now but it can be changed easily enough.

Next step is generate a PDF export for the report.