Update on 11 April 14:48
The engineers version is now exporting reports correctly. It has the ability to look in a specific folder and then export the PDF to that folder. This folder must have a resourcesEng folder which is where the pngs go. I am thinking of relocating this to an "install" folder.

the UI has been updated to grey out the buttons when they aren't required. This forces the user to perform an import first when opening the software. The buttons do not grey out again until the software is restarted. This is a matter of management preference as to whether this should carry on like this or not.

The user can switch between analysers without closing the software by selecting import again and reimporting a file in a different folder. The software will switch to this folder and export the PDF to that folder.