Update on 13 April 16:38
Calibrator version is coming along nicely.

The export to email function works correctly and the UI is mostly sorted out.

Currently to only issue is the loading of the Historical data. I haven't quite got the logic sorted. I am currently pulling the previous months json file, if the analyser is older than 2 months. This data loads as normal which means the data is off by a month. This will have to be corrected somehow.