Update on 27 April 13:58
Export report Email text body should be able to update based on what analyser is being exported. This should pull in the engineer's email address and update the body text automatically.

Attempting to implement the auto update of the psa checklist. This will be fairly rough updating and wont check if the list is open in multiple locations but will minimise the time that it is open.