Update on 27 April 16:00
Export report Email text body should be able to update based on what analyser is being exported. This should pull in the engineer's email address and update the body text automatically.

Attempting to implement the auto update of the psa checklist. This will be fairly rough updating and wont check if the list is open in multiple locations but will minimise the time that it is open.
There is a method there but it still requires some debugging. It should work but it has not been properly tested.

importing historical data is still an issue. I am thinking a seperate function will be required to import it into a different spot.