Update on 28 April 16:00
Export report Email text body should be able to update based on what analyser is being exported. This should pull in the engineer's email address and update the body text automatically.

The software now autoupdates the PSA Checklist properly. It is still hardcoded to access the C:\PSAGen location until release as it will overwrite the file as soon as it starts running in the wild.

importing historical data is still an issue. I am thinking a seperate function will be required to import it into a different spot.