Update on 29 April 09:25
Export report Email text body and subject updates automatically based on the analyser and engineer in question. It does this by accessing a PSA Report config excel document. This is currently in the PSAGen folder but this will need to be changed to get it from the server location once ready for release.

The software now autoupdates the PSA Checklist properly. It is still hardcoded to access the C:\PSAGen location until release as it will overwrite the file as soon as it starts running in the wild.

importing historical data is still an issue. I am thinking a seperate function will be required to import it into a different spot.

The Page 1 unhide when previewing should only take place after the import of last month's data.

The import of last month's data should be selectable.