Update on 29 April 12:54
Export report Email text body and subject updates automatically based on the analyser and engineer in question. It does this by accessing a PSA Report config excel document. This is currently in the PSAGen folder but this will need to be changed to get it from the server location once ready for release.

The software now autoupdates the PSA Checklist properly. It is still hardcoded to access the C:\PSAGen location until release as it will overwrite the file as soon as it starts running in the wild.

importing historical data issue has been resolved.

The Page 1 unhides after last month's data has been selected or the default used.

The import of last month's data is now selectable selectable.

Resolve the chart naming issue. it still only records the result number not the element name.