Update on 3 May 2022 09:12
Blank report generation is now working along with variable email outputs and subject lines.

Export report Email text body and subject updates automatically based on the analyser and engineer in question. It does this by accessing a PSA Report config Excel document. This is currently in the PSAGen folder but this will need to be changed to get it from the server location once ready for release.

The software now auto-updates the PSA Checklist properly. It is still hardcoded to access the C:\PSAGen location until release as it will overwrite the file as soon as it starts running in the wild.

importing historical data issue has been resolved.

The Page 1 un-hides after last month's data has been selected or the default used.

The import of last month's data is now selectable.

Resolve the chart naming issue has been sorted by using the master config Excel document.

The analyser list on Page 1 now only populates with analyser which have PSA report Required set to Yes in the PSA Report config file.