Update on 5 May 2022 16:12
Demoed to R&D Group and they are happy but want Lucas's suggestions implemented.

I have to resolve the resolution issue as when the resolution of the screen changes the program goes to hell.

The plotting of the detector stability plot is still and issue and the x axis needs work.

Packaging of Calibrator version now works. Make sure to use the batch file method

Page control has also been added to all pages of the calibrator version.

The page control for each page still needs to be added to the engineers version.