Update on 6 May 2022 14:55
I have to resolve the resolution issue as when the resolution of the screen changes the program goes to hell.

Detector stability plot has been sorted out. The DPI of all plot increased to 600dpi.

Packaging of Calibrator version now works. Make sure to use the batch file method

The page control for each page still needs to be added to the engineers version.